\name{rmse_location_yourcats}
\alias{rmse_location_yourcats}

\title{Get in and out of sample RMSE for each location}

\description{yourcats class summary function}

\usage{
\method{rmse_yourcats}{yourcast}(object)
}

\arguments{
    \item{object}{\code{\link{yourcast}} output object to be summarized}
}

\value{A data.frame with in/out of sample rmse for all locations.
}

\author{Neal Marquez \email{nmarquez@uw.edu}}

\seealso{\code{\link{yourcast}},\code{\link{print.summary.yourcast}} }
