\name{rmse_yourcats}
\alias{rmse_yourcats}

\title{Get in and out of sample RMSE}

\description{yourcats class summary function}

\usage{
\method{rmse_yourcats}{yourcast}(object)
}

\arguments{
    \item{object}{\code{\link{yourcast}} output object to be summarized}
}

\value{A vector of length 2 with out and in RMSE
}

\author{Neal Marquez \email{nmarquez@uw.edu}}

\seealso{\code{\link{yourcast}},\code{\link{print.summary.yourcast}} }